tests are passing, corrected Map Logic (all match in shapeHandler)
