+++ app/src/main/java/hexlet/code/Validator.java
@@ -15,7 +15,7 @@ public class Validator {
     @Getter
     private List<BaseSchema> schemas;
 
-    Validator() {
+    public Validator() {
         this.schemas = new ArrayList<>();
     }
