+++ app/src/main/java/hexlet/code/validators/MapSchema.java
@@ -3,10 +3,13 @@ package hexlet.code.validators;
 import hexlet.code.Validator;
 import lombok.Setter;
 
+import java.util.HashMap;
 import java.util.Map;
 
 public class MapSchema<T> extends BaseSchema<Map<T, T>> {
     private int sizeOf;
+    private Map<String, StringSchema> stringSchemas = new HashMap<>();
+    private Map<Integer, NumberSchema> numberSchemas = new HashMap<>();
 
     @Setter
     private Validator validator;
@@ -15,7 +18,7 @@ public class MapSchema<T> extends BaseSchema<Map<T, T>> {
         super();
     }
 
-    public BaseSchema required() {
+    public MapSchema required() {
         this.getInternalState().add("required");
         return this;
     }
@@ -26,20 +29,85 @@ public class MapSchema<T> extends BaseSchema<Map<T, T>> {
                 .allMatch(field -> stateHandler(field, value));
     }
 
-    public void sizeof(int size) {
+    public MapSchema sizeof(int size) {
         this.sizeOf = size;
         this.getInternalState().add("sizeof");
+        return this;
     }
 
     private boolean stateHandler(String field, Map<T, T> value) {
         return switch (field) {
             case "required" -> value != null;
             case "sizeof" -> value.size() >= this.sizeOf;
-            default -> false;
+            case "shape" -> shapeHandler(value);
+            default -> throw new RuntimeException("There is no settings for the schema");
         };
     }
 
-    public void shape(Map<T, BaseSchema<T>> schemas) {
+    public MapSchema shape(Map<T, BaseSchema<T>> schemas) {
+        this.getInternalState().add("shape");
+        schemas.forEach((key, value) -> {
+            if (key instanceof String) {
+                this.stringSchemas.put(key.toString(), (StringSchema) value);
+            } else if (key instanceof Integer) {
+                this.numberSchemas.put((Integer) key, (NumberSchema) value);
+            } else {
+                throw new RuntimeException("Unknown instance");
+            }
+        });
+        return this;
+    }
 
+
+    public boolean shapeHandler(Map<T, T> value) {
+        return value.entrySet().stream()
+                .allMatch(entry -> {
+                    boolean result = false;
+                    if (entry.getKey() instanceof String) {
+                        Map<String, String> valueString = new HashMap<>();
+                        valueString.put(entry.getKey().toString(),
+                                entry.getValue() == null ? null : entry.getValue().toString());
+                        result = shapeStringHandler(valueString);
+                    } else if (entry.getKey() instanceof Integer) {
+                        Map<Integer, Integer> valueInt = new HashMap<>();
+                        valueInt.put((Integer)entry.getKey(),
+                                entry.getValue() == null ? null : (Integer)entry.getValue());
+                        result = shapeIntegerHandler(valueInt);
+                    }
+                    return result;
+                });
+    }
+//        for (Map.Entry<T, T> entry : value.entrySet()) {
+//            if (entry.getKey() instanceof String) {
+//                Map<String, String> valueString = new HashMap<>();
+//                valueString.put(entry.getKey().toString(),
+//                        entry.getValue() == null ? null : entry.getValue().toString());
+//                return shapeStringHandler(valueString);
+//            } else if (entry.getKey() instanceof Integer) {
+//                Map<Integer, Integer> valueInt = new HashMap<>();
+//                valueInt.put((Integer)entry.getKey(),
+//                        entry.getValue() == null ? null : (Integer)entry.getValue());
+//                return shapeIntegerHandler(valueInt);
+//            }
+//        }
+//        return false;
+//    }
+
+    public boolean shapeStringHandler(Map<String, String> value) {
+        for (Map.Entry<String, String> entry : value.entrySet()) {
+            String key = entry.getKey();
+            StringSchema check = this.stringSchemas.getOrDefault(key, null);
+            return check.isValid(entry.getValue());
+        }
+        return true;
+    }
+
+    public boolean shapeIntegerHandler(Map<Integer, Integer> value) {
+        for (Map.Entry<Integer, Integer> entry : value.entrySet()) {
+            Integer key = entry.getKey();
+            NumberSchema check = this.numberSchemas.getOrDefault(key, null);
+            return check.isValid(entry.getValue());
+        }
+        return true;
     }
 }
